Add SearchPrecision class to configure search accuracy
Introduced a new class `SearchPrecision` in the `CutCraftEngineDataInput.Configuration` namespace to specify search accuracy. The class inherits from `DataGroupInput` and includes two properties: `firstLevel` and `nextLevels`, with default values set for both. A constructor is provided for initialization. Added using directives for necessary namespaces. Commented-out validation check for `nextLevels` to be added later.