Add DeviceRepository class with Device initialization
Introduced a new internal class `DeviceRepository` in the `CutCraftEngineData.DataInput` namespace. Added necessary `using` directives. The class includes a public property `device`, which is a list of `Device` objects. The constructor initializes the `device` list and adds a `Device` object via the `DeviceSektor450` method. The `DeviceSektor450` method sets various properties for the `Device` object.